se solucionan errores en vista details

+++ database/migrations/2020_06_08_195436_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->id();
             $table->string('client');
             $table->string('address');
-            $table->decimal('cost');
+            $table->integer('cost');
             $table->date('deliver_date');
             $table->time('deliver_hour');
             $table->boolean('enable')->default(1);
